Updated and tested boggle
Based on new lessons created by RD in Aug 2017. The lesson is structured so that the student follows a lesson unit, and then is provided the working solution code in the next lesson unit. As a result, we only need to include full solution code for lessons and challenges in github (as opposed to segmented structure we had before)